Server runs with no errors lots of debugging still in place. Testing another player connecting seems to be the next logical step though will do some testing before that.
